backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pgvector
pydantic
pydantic-settings
python-dotenv
openai
